RSI Results
These are the results of our RSI backtests on the BTCUSDT pair on the most common timeframes. 
All tests were conducted with the amount of transaction in 100$, without leverage. This is done for the convenience of your calculations, as 100$ = 100%.  
During the testing, we used dynamic RSI parameters, they are shown in the RSI_limit column. As we can see RSI parameters have a very strong influence on the number of trades. The file also contains: 
Maximum drawdown of a trade
Maximum deposit drawdown 
Maximum gain in a deal 
Maximum deposit growth
Remember, this is not financial advice, but as you can see, even RSI has a high winrate and can be profitable.